Add files via upload
This is the notebook for the Customer Churn Project. it was done on google collabs and is compatible with jyupiter notebooks. In case of any questions, you can reach me via twitter or instagram @datagulf1 or via my email  derrickkuria44@gmail.com. If you find it helpful , star the repository and follow me. I follow back.